smartclassroom/backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
django-filter==24.3
channels==4.1.0
channels-redis==4.2.0
dj-database-url==2.3.0
requests==2.32.3
psycopg2-binary==2.9.11
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
